refactor: Remove GQL finish step for Create policy component
Also removes the usePolicy mutation hook.
