+++ src/SmartComponents/CreatePolicy/FinishedCreatePolicy/FinishedCreatePolicy.js
@@ -1,21 +1,142 @@
-import React from 'react';
-import useAPIV2FeatureFlag from '../../../Utilities/hooks/useAPIV2FeatureFlag';
-import { Bullseye, Spinner } from '@patternfly/react-core';
-import FinishedCreatePolicyRest from './FinishedCreatePolicyRest';
-import FinishedCreatePolicyGraphQL from './FinishedCreatePolicyGraphQL';
+import React, { useCallback } from 'react';
+import FinishedCreatePolicyBase from './FinishedCreatePolicyBase';
+import useCreatePolicy from '../../../Utilities/hooks/api/useCreatePolicy';
+import useAssignRules from '../../../Utilities/hooks/api/useAssignRules';
+import useAssignSystems from '../../../Utilities/hooks/api/useAssignSystems';
+import useTailorings from '../../../Utilities/hooks/api/useTailorings';
+import useUpdateTailoring from '../../../Utilities/hooks/api/useUpdateTailoring';
+
+export const useUpdatePolicy = () => {
+  const { fetch: createPolicy } = useCreatePolicy({ skip: true });
+  const { fetch: assignRules } = useAssignRules({ skip: true });
+  const { fetch: assignSystems } = useAssignSystems({ skip: true });
+  const { fetch: fetchTailorings } = useTailorings({ skip: true });
+  const { fetch: updateTailoring } = useUpdateTailoring({ skip: true }); // to update value overrides
+
+  const updatedPolicy = useCallback(
+    async (
+      _,
+      {
+        name,
+        description,
+        businessObjective,
+        complianceThreshold,
+        benchmarkId,
+        selectedRuleRefIds,
+        hosts,
+        valueOverrides,
+      },
+      onProgress
+    ) => {
+      const minorVersions = selectedRuleRefIds.map(
+        ({ osMinorVersion }) => osMinorVersion
+      );
+      const expectedUpdates =
+        3 + minorVersions.length + Object.keys(valueOverrides).length;
+      let progress = 0;
+
+      const dispatchProgress = () => {
+        if (onProgress) {
+          onProgress(++progress / expectedUpdates);
+        }
+      };
+
+      const createPolicyResponse = await createPolicy(
+        [
+          undefined,
+          {
+            title: name,
+            description,
+            business_objective: businessObjective.title,
+            compliance_threshold: complianceThreshold,
+            profile_id: benchmarkId,
+          },
+        ],
+        false
+      );
+
+      dispatchProgress();
+
+      const { id: newPolicyId } = createPolicyResponse.data;
+
+      await assignSystems(
+        [newPolicyId, undefined, { ids: hosts.map(({ id }) => id) }],
+        false
+      );
+
+      dispatchProgress();
+
+      const fetchPolicyResponse = await fetchTailorings(
+        [
+          newPolicyId,
+          undefined,
+          100 /** to ensure we fetch all tailorings at once */,
+        ],
+        false
+      );
+      const tailorings = fetchPolicyResponse.data;
+
+      dispatchProgress();
+
+      for (const tailoring of tailorings) {
+        const rulesToAssign = selectedRuleRefIds.find(
+          ({ osMinorVersion }) =>
+            Number(osMinorVersion) === tailoring.os_minor_version
+        ).ruleRefIds;
+
+        await assignRules(
+          [
+            newPolicyId,
+            tailoring.id,
+            undefined,
+            {
+              ids: rulesToAssign,
+            },
+          ],
+          false
+        );
+
+        dispatchProgress();
+      }
+
+      for (const tailoring of tailorings) {
+        const tailoringValueOverrides =
+          valueOverrides[tailoring.os_minor_version];
+
+        if (
+          tailoringValueOverrides === undefined ||
+          Object.keys(tailoringValueOverrides).length === 0
+        ) {
+          continue;
+        }
+
+        await updateTailoring(
+          [
+            newPolicyId,
+            tailoring.id,
+            undefined,
+            {
+              value_overrides: tailoringValueOverrides,
+            },
+          ],
+          false
+        );
+
+        dispatchProgress();
+      }
+
+      return { id: newPolicyId };
+    },
+    [createPolicy, assignSystems, fetchTailorings, assignRules, updateTailoring]
+  );
+
+  return updatedPolicy;
+};
 
 const FinishedCreatePolicy = (props) => {
-  const apiV2Enabled = useAPIV2FeatureFlag();
+  const updatePolicy = useUpdatePolicy();
 
-  return apiV2Enabled === undefined ? (
-    <Bullseye>
-      <Spinner />
-    </Bullseye>
-  ) : apiV2Enabled ? (
-    <FinishedCreatePolicyRest {...props} />
-  ) : (
-    <FinishedCreatePolicyGraphQL {...props} />
-  );
+  return <FinishedCreatePolicyBase updatePolicy={updatePolicy} {...props} />;
 };
 
 export default FinishedCreatePolicy;
