+++ src/SmartComponents/EditPolicy/hooks.test.js
@@ -1,9 +1,6 @@
 import { renderHook, act, waitFor } from '@testing-library/react';
 import { useOnSave } from './hooks';
 
-import { usePolicy } from 'Mutations';
-jest.mock('Mutations');
-
 import { dispatchNotification } from 'Utilities/Dispatcher';
 jest.mock('Utilities/Dispatcher');
 
@@ -11,9 +8,6 @@ jest.mock('Utilities/hooks/useAnchor', () => ({
   __esModule: true,
   default: () => () => ({}),
 }));
-
-import useAPIV2FeatureFlag from '../../Utilities/hooks/useAPIV2FeatureFlag';
-jest.mock('../../Utilities/hooks/useAPIV2FeatureFlag');
 
 describe('useOnSave', function () {
   const policy = {};
@@ -26,13 +20,9 @@ describe('useOnSave', function () {
     onSaveCallBack.mockReset();
     onErrorCallback.mockReset();
     dispatchNotification.mockImplementation(mockedNotification);
-    useAPIV2FeatureFlag.mockImplementation(() => false);
   });
 
   it('returns a function to call with a policy and updated policy', async () => {
-    usePolicy.mockImplementation(() => {
-      return () => Promise.resolve();
-    });
     const { result } = renderHook(() =>
       useOnSave(policy, updatedPolicy, {
         onSave: onSaveCallBack,
@@ -58,9 +48,6 @@ describe('useOnSave', function () {
   });
 
   it('returns a function to call with a policy and updated policy and can raise an error', async () => {
-    usePolicy.mockImplementation(() => {
-      return () => Promise.reject({});
-    });
     const { result } = renderHook(() =>
       useOnSave(policy, updatedPolicy, {
         onSave: onSaveCallBack,
